pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "momentum-signal-engine"
version = "0.1.0"
description = "Stock trading analysis platform for high-probability momentum setups"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "alpaca-py>=0.28.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "ta>=0.11.0",
    "typer>=0.12.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "rich>=13.8.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[project.scripts]
mse = "src.cli.commands:app"

[tool.setuptools.packages.find]
include = ["src*", "config*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
